requirements.txt
opencv-python
socketIO-client==0.7.2
websocket-client==1.2.1
python-socketio==5.4.0
Flask
Flask-Cors
prettytable
urllib3
requests
easyocr
pytesseract
numpy
pywin32==300
pywin32-ctypes==0.2.0
pywinpty==0.5.7